api/requirements.txt
djangorestframework == 3.14.0
django-cors-headers
pandas
openpyxl
sqlalchemy
mysqlclient
statsmodels
numpy
pmdarima
scikit-learn
plotly